+++ src/fields/Address.php
@@ -1,29 +1,44 @@
 <?php
+/**
+ * @link https://sprout.barrelstrengthdesign.com
+ * @copyright Copyright (c) Barrel Strength Design LLC
+ * @license https://craftcms.github.io/license
+ */
 
 namespace barrelstrength\sproutfields\fields;
 
 use barrelstrength\sproutbasefields\base\AddressFieldTrait;
-use CommerceGuys\Addressing\AddressFormat\AddressFormatRepository;
-use CommerceGuys\Addressing\Formatter\DefaultFormatter;
+use barrelstrength\sproutbasefields\models\Address as AddressModel;
+use barrelstrength\sproutbasefields\SproutBaseFields;
 use CommerceGuys\Addressing\Address as CommerceGuysAddress;
-use CommerceGuys\Addressing\Subdivision\SubdivisionRepository;
+use CommerceGuys\Addressing\AddressFormat\AddressFormatRepository;
 use CommerceGuys\Addressing\Country\CountryRepository;
+use CommerceGuys\Addressing\Formatter\DefaultFormatter;
+use CommerceGuys\Addressing\Subdivision\SubdivisionRepository;
 use GraphQL\Type\Definition\Type;
 use GraphQL\Type\Definition\ObjectType;
+use Craft;
 use craft\base\ElementInterface;
 use craft\base\Field;
 use craft\base\PreviewableFieldInterface;
 use craft\gql\GqlEntityRegistry;
 use craft\gql\TypeLoader;
-use Craft;
-use yii\db\Schema;
+use craft\errors\SiteNotFoundException;
+use Throwable;
+use Twig\Error\LoaderError;
+use Twig\Error\RuntimeError;
+use Twig\Error\SyntaxError;
+use yii\db\Exception;
+use yii\db\StaleObjectException;
 
 /**
  * Class Address
  *
  * @package barrelstrength\sproutfields\fields
  *
- * @property string $contentColumnType
+ * @property array       $elementValidationRules
+ * @property null|string $settingsHtml
+ * @property string      $contentColumnType
  */
 class Address extends Field implements PreviewableFieldInterface
 {
@@ -33,6 +48,17 @@ class Address extends Field implements PreviewableFieldInterface
      * @var string|null
      */
     public $value;
+
+    public static function supportedTranslationMethods(): array
+    {
+        return [
+            self::TRANSLATION_METHOD_NONE,
+            self::TRANSLATION_METHOD_SITE,
+            self::TRANSLATION_METHOD_SITE_GROUP,
+            self::TRANSLATION_METHOD_LANGUAGE,
+            self::TRANSLATION_METHOD_CUSTOM,
+        ];
+    }
 
     /**
      * @inheritdoc
@@ -45,9 +71,73 @@ class Address extends Field implements PreviewableFieldInterface
     /**
      * @inheritdoc
      */
-    public function getContentColumnType(): string
+    public static function hasContentColumn(): bool
     {
-        return Schema::TYPE_INTEGER;
+        return false;
+    }
+
+    /**
+     * @return string|null
+     * @throws LoaderError
+     * @throws RuntimeError
+     * @throws SyntaxError
+     * @throws SiteNotFoundException
+     */
+    public function getSettingsHtml()
+    {
+        return SproutBaseFields::$app->addressField->getSettingsHtml($this);
+    }
+
+    /**
+     * @param                       $value
+     * @param ElementInterface|null $element
+     *
+     * @return string
+     * @throws LoaderError
+     * @throws RuntimeError
+     * @throws SyntaxError
+     */
+    public function getInputHtml($value, ElementInterface $element = null): string
+    {
+        return SproutBaseFields::$app->addressField->getInputHtml($this, $value, $element);
+    }
+
+    /**
+     * How the field output will appear for Revisions
+     *
+     * @param                  $value
+     * @param ElementInterface $element
+     *
+     * @return string
+     */
+    public function getStaticHtml($value, ElementInterface $element): string
+    {
+        return SproutBaseFields::$app->addressField->getStaticHtml($this, $value, $element);
+    }
+
+    /**
+     * @param                       $value
+     * @param ElementInterface|null $element
+     *
+     * @return AddressModel|mixed|null
+     */
+    public function normalizeValue($value, ElementInterface $element = null)
+    {
+        return SproutBaseFields::$app->addressField->normalizeValue($this, $value, $element);
+    }
+
+    /**
+     * @param ElementInterface $element
+     * @param bool             $isNew
+     *
+     * @throws Throwable
+     * @throws Exception
+     * @throws StaleObjectException
+     */
+    public function afterElementSave(ElementInterface $element, bool $isNew)
+    {
+        SproutBaseFields::$app->addressField->afterElementSave($this, $element, $isNew);
+        parent::afterElementSave($element, $isNew);
     }
 
     /**
@@ -55,6 +145,7 @@ class Address extends Field implements PreviewableFieldInterface
      */
     public function getTableAttributeHtml($value, ElementInterface $element): string
     {
+
         if (!$value) {
             return '';
         }
@@ -105,5 +196,4 @@ class Address extends Field implements PreviewableFieldInterface
 
         return $addressType;
     }
-
 }