+++ src/fields/Phone.php
@@ -1,21 +1,23 @@
 <?php
+/**
+ * @link https://sprout.barrelstrengthdesign.com
+ * @copyright Copyright (c) Barrel Strength Design LLC
+ * @license https://craftcms.github.io/license
+ */
 
 namespace barrelstrength\sproutfields\fields;
 
+use barrelstrength\sproutbasefields\models\Phone as PhoneModel;
 use barrelstrength\sproutbasefields\SproutBaseFields;
 use Craft;
 use craft\base\Element;
 use craft\base\ElementInterface;
 use craft\base\Field;
 use craft\base\PreviewableFieldInterface;
-use libphonenumber\PhoneNumberUtil;
 use Twig\Error\LoaderError;
 use Twig\Error\RuntimeError;
 use Twig\Error\SyntaxError;
 use yii\db\Schema;
-
-use CommerceGuys\Intl\Country\CountryRepository;
-use barrelstrength\sproutbasefields\models\Phone as PhoneModel;
 
 /**
  *
@@ -50,7 +52,9 @@ class Phone extends Field implements PreviewableFieldInterface
      * @var string|null
      */
     public $customPatternToggle;
+
     public $mask;
+
     public $inputMask;
 
     public static function displayName(): string
@@ -67,6 +71,31 @@ class Phone extends Field implements PreviewableFieldInterface
     }
 
     /**
+     * @param                       $value
+     * @param ElementInterface|null $element
+     *
+     * @return PhoneModel|mixed|null
+     */
+    public function normalizeValue($value, ElementInterface $element = null)
+    {
+        return SproutBaseFields::$app->phoneField->normalizeValue($this, $value, $element);
+    }
+
+    /**
+     * @param                       $value
+     *
+     * @param ElementInterface|null $element
+     *
+     * @return array|mixed|string|null
+     */
+    public function serializeValue($value, ElementInterface $element = null)
+    {
+        $value = SproutBaseFields::$app->phoneField->serializeValue($value);
+
+        return parent::serializeValue($value, $element);
+    }
+
+    /**
      * @inheritdoc
      *
      * @throws LoaderError
@@ -75,12 +104,7 @@ class Phone extends Field implements PreviewableFieldInterface
      */
     public function getSettingsHtml()
     {
-        return Craft::$app->getView()->renderTemplate(
-            'sprout-base-fields/_components/fields/formfields/phone/settings',
-            [
-                'field' => $this,
-            ]
-        );
+        return SproutBaseFields::$app->phoneField->getSettingsHtml($this);
     }
 
     /**
@@ -96,95 +120,7 @@ class Phone extends Field implements PreviewableFieldInterface
      */
     public function getInputHtml($value, ElementInterface $element = null): string
     {
-        $name = $this->handle;
-        $countryId = Craft::$app->getView()->formatInputId($name.'-country');
-        $inputId = Craft::$app->getView()->formatInputId($name);
-        $namespaceInputId = Craft::$app->getView()->namespaceInputId($inputId);
-        $namespaceCountryId = Craft::$app->getView()->namespaceInputId($countryId);
-        $countries = $this->getCountries();
-
-        $country = $value['country'] ?? $this->country;
-        $val = $value['phone'] ?? null;
-
-        return Craft::$app->getView()->renderTemplate(
-            'sprout-base-fields/_components/fields/formfields/phone/input',
-            [
-                'namespaceInputId' => $namespaceInputId,
-                'namespaceCountryId' => $namespaceCountryId,
-                'id' => $inputId,
-                'countryId' => $countryId,
-                'name' => $this->handle,
-                'value' => $val,
-                'placeholder' => $this->placeholder,
-                'countries' => $countries,
-                'country' => $country,
-                'limitToSingleCountry' => $this->limitToSingleCountry
-            ]
-        );
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function normalizeValue($value, ElementInterface $element = null)
-    {
-        $phoneInfo = [];
-
-        if (is_array($value) && $element) {
-            $namespace = $element->getFieldParamNamespace();
-            $namespace = $namespace.'.'.$this->handle;
-            $phoneInfo = Craft::$app->getRequest()->getBodyParam($namespace);
-            // bad phone or empty phone
-        }
-
-        if (is_string($value)) {
-            $phoneInfo = json_decode($value, true);
-        }
-
-        if (!isset($phoneInfo['phone'], $phoneInfo['country'])) {
-            return $value;
-        }
-
-        // Always return array
-        return new PhoneModel($phoneInfo['phone'], $phoneInfo['country']);
-    }
-
-    /**
-     * @param mixed                 $value
-     * @param ElementInterface|null $element
-     *
-     * @return array|mixed|null|string
-     */
-    public function serializeValue($value, ElementInterface $element = null)
-    {
-        // Submitting an Element to be saved
-        if (is_object($value) && get_class($value) == PhoneModel::class) {
-            return $value->getAsJson();
-        }
-
-        // Save the phone as json with the number and country
-        return $value;
-    }
-
-    public function getCountries()
-    {
-        $phoneUtil = PhoneNumberUtil::getInstance();
-        $regions = $phoneUtil->getSupportedRegions();
-        $countries = [];
-
-        foreach ($regions as $countryCode) {
-            $code = $phoneUtil->getCountryCodeForRegion($countryCode);
-            $countryRepository = new CountryRepository;
-            $country = $countryRepository->get($countryCode);
-
-            if ($country) {
-                $countries[$countryCode] = $country->getName().' +'.$code;
-            }
-        }
-
-        asort($countries);
-
-        return $countries;
+        return SproutBaseFields::$app->phoneField->getInputHtml($this, $value);
     }
 
     /**
@@ -211,20 +147,16 @@ class Phone extends Field implements PreviewableFieldInterface
     {
         $value = $element->getFieldValue($this->handle);
 
-        if ($this->required && !$value->phone) {
-            $element->addError(
-                $this->handle,
-                Craft::t('sprout-fields', '{field} cannot be blank', [
-                    'field' => $this->name
-                ])
-            );
+        // Don't run validation if a field is not required and has no value for the phone number
+        if (!$this->required && empty($value->phone)) {
+            return;
         }
 
-        if ($value->country && $value->phone && !SproutBaseFields::$app->phoneField->validate($value->phone, $value->country)) {
-            $element->addError(
-                $this->handle,
-                SproutBaseFields::$app->phoneField->getErrorMessage($this, $value->country)
-            );
+        $isValid = SproutBaseFields::$app->phoneField->validate($value);
+
+        if (!$isValid) {
+            $message = SproutBaseFields::$app->phoneField->getErrorMessage($this, $value->country);
+            $element->addError($this->handle, $message);
         }
     }