+++ src/migrations/m180221_161522_notes_fields.php
@@ -1,9 +1,14 @@
 <?php
+/**
+ * @link https://sprout.barrelstrengthdesign.com
+ * @copyright Copyright (c) Barrel Strength Design LLC
+ * @license https://craftcms.github.io/license
+ */
 
 namespace barrelstrength\sproutfields\migrations;
 
-use craft\db\Migration;
 use barrelstrength\sproutfields\fields\Notes;
+use craft\db\Migration;
 use craft\db\Query;
 use craft\helpers\Json;
 
@@ -33,6 +38,7 @@ class m180221_161522_notes_fields extends Migration
 
             $this->update('{{%fields}}', ['type' => Notes::class, 'settings' => $settingsAsJson], ['id' => $noteField['id']], [], false);
         }
+
         // end notes
         return true;
     }
@@ -43,6 +49,7 @@ class m180221_161522_notes_fields extends Migration
     public function safeDown(): bool
     {
         echo "m180221_161522_notes_fields cannot be reverted.\n";
+
         return false;
     }
 }