+++ src/migrations/m180221_161526_link_fields.php
@@ -1,9 +1,14 @@
 <?php
+/**
+ * @link https://sprout.barrelstrengthdesign.com
+ * @copyright Copyright (c) Barrel Strength Design LLC
+ * @license https://craftcms.github.io/license
+ */
 
 namespace barrelstrength\sproutfields\migrations;
 
-use craft\db\Migration;
 use barrelstrength\sproutfields\fields\Url;
+use craft\db\Migration;
 
 /**
  * m180221_161526_link_fields migration.
@@ -31,6 +36,7 @@ class m180221_161526_link_fields extends Migration
     public function safeDown(): bool
     {
         echo "m180221_161526_link_fields cannot be reverted.\n";
+
         return false;
     }
 }