+++ src/migrations/m190524_000000_add_dynamic_rendering_and_standalone_predefined_date_field_sproutfields.php
@@ -1,8 +1,11 @@
-<?php /** @noinspection ClassConstantCanBeUsedInspection */
+<?php /**
+ * @link https://sprout.barrelstrengthdesign.com
+ * @copyright Copyright (c) Barrel Strength Design LLC
+ * @license https://craftcms.github.io/license
+ */ /** @noinspection ClassConstantCanBeUsedInspection */
 
 namespace barrelstrength\sproutfields\migrations;
 
-use barrelstrength\sproutbasefields\migrations\m190521_000000_add_predefined_data_type_setting;
 use barrelstrength\sproutbasefields\migrations\m190524_000000_add_dynamic_rendering_and_standalone_predefined_date_field;
 use craft\db\Migration;
 
@@ -31,6 +34,7 @@ class m190524_000000_add_dynamic_rendering_and_standalone_predefined_date_field_
     public function safeDown(): bool
     {
         echo "m190524_000000_add_dynamic_rendering_and_standalone_predefined_date_field_sproutfields cannot be reverted.\n";
+
         return false;
     }
 }